refactor(portfolio): split PortfolioService into pricing/read/write/stop collaborators behind a facade
